DFE-708 Split MultiHeaderTable out
