+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/FixedHeaderGroupedDataTable.module.scss
@@ -1,7 +1,5 @@
 @import '~explore-education-statistics-common/src/styles/govuk-base';
 
-$border-width: 4px;
-
 .container {
   max-height: 80vh;
   overflow: scroll;
@@ -44,28 +42,4 @@ $border-width: 4px;
   top: 0;
   width: auto;
   z-index: 4;
-}
-
-.intersectionCell {
-  border-bottom: 0 !important;
-}
-
-.figure {
-  margin: 0;
-
-  figcaption {
-    padding-bottom: govuk-spacing(4);
-  }
-}
-
-.borderLeft {
-  border-left: $border-width solid govuk-colour('black') !important;
-}
-
-.borderRight {
-  border-right: $border-width solid govuk-colour('black') !important;
-}
-
-.borderBottom {
-  border-bottom: $border-width solid govuk-colour('black') !important;
 }
