+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/FixedHeaderGroupedDataTable.tsx
@@ -1,19 +1,10 @@
-import classNames from 'classnames';
 import throttle from 'lodash/throttle';
-import times from 'lodash/times';
 import React, { forwardRef, Ref, useEffect, useRef } from 'react';
 import DataTableKeys from './DataTableKeys';
 import styles from './FixedHeaderGroupedDataTable.module.scss';
+import MultiHeaderTable, { RowGroup } from './MultiHeaderTable';
 
 const dataTableCaption = 'dataTableCaption';
-
-export interface RowGroup {
-  label: string;
-  rows: {
-    label: string;
-    columnGroups: string[][];
-  }[];
-}
 
 interface Props {
   caption: string;
@@ -21,154 +12,6 @@ interface Props {
   innerRef?: Ref<HTMLElement>;
   rowGroups: RowGroup[];
 }
-
-type InnerTableProps = {
-  ariaHidden?: boolean;
-  className?: string;
-  isStickyHeader?: boolean;
-  isStickyColumn?: boolean;
-} & Props;
-
-const GroupedDataTable = forwardRef<HTMLTableElement, InnerTableProps>(
-  (
-    {
-      ariaHidden,
-      className,
-      headers,
-      isStickyColumn,
-      isStickyHeader,
-      rowGroups,
-    },
-    ref,
-  ) => {
-    const maxColSpan = headers.reduce(
-      (total, header) => total * header.length,
-      1,
-    );
-
-    const getColSpan = (headerRowIndex: number) => {
-      return headers
-        .slice(headerRowIndex + 1)
-        .reduce((total, row) => total * row.length, 1);
-    };
-
-    return (
-      <table
-        aria-hidden={ariaHidden}
-        aria-labelledby={dataTableCaption}
-        className={classNames('govuk-table', className)}
-        ref={ref}
-      >
-        <thead>
-          {headers.map((columns, rowIndex) => {
-            const colSpan = getColSpan(rowIndex) || 1;
-            const isColGroup = rowIndex !== headers.length - 1;
-
-            return (
-              // eslint-disable-next-line react/no-array-index-key
-              <tr key={rowIndex}>
-                {rowIndex === 0 && (
-                  <th
-                    colSpan={2}
-                    rowSpan={headers.length}
-                    className={classNames(
-                      styles.borderBottom,
-                      styles.borderRight,
-                    )}
-                  />
-                )}
-
-                {!isStickyColumn &&
-                  times(maxColSpan / (colSpan * columns.length), () =>
-                    columns.map((column, columnIndex) => {
-                      const key = `${column}_${columnIndex}`;
-
-                      return (
-                        <th
-                          className={classNames({
-                            'govuk-table__header--numeric': !isColGroup,
-                            'govuk-table__header--center': isColGroup,
-                            [styles.borderBottom]: !isColGroup,
-                            [styles.borderLeft]:
-                              columnIndex === 0 || isColGroup,
-                          })}
-                          colSpan={colSpan}
-                          scope={colSpan > 1 ? 'colgroup' : 'col'}
-                          key={key}
-                        >
-                          {column}
-                        </th>
-                      );
-                    }),
-                  )}
-              </tr>
-            );
-          })}
-        </thead>
-
-        {rowGroups.map((group, groupIndex) => {
-          const groupKey = `${group.label}-${groupIndex}`;
-
-          return (
-            !isStickyHeader && (
-              <tbody key={groupKey} className={styles.borderBottom}>
-                {group.rows.map((row, rowIndex) => {
-                  const rowKey = `${groupKey}_${row.label}_${rowIndex}`;
-
-                  return (
-                    <tr key={rowKey}>
-                      {rowIndex === 0 && (
-                        <th
-                          scope="rowgroup"
-                          rowSpan={group.rows.length || 1}
-                          className={classNames(styles.borderBottom)}
-                        >
-                          {group.label}
-                        </th>
-                      )}
-                      <th
-                        scope="row"
-                        className={classNames(
-                          'govuk-table__cell--numeric',
-                          styles.borderRight,
-                        )}
-                      >
-                        {row.label}
-                      </th>
-                      {!isStickyColumn &&
-                        row.columnGroups.flatMap((colGroup, colGroupIndex) =>
-                          colGroup.map((column, columnIndex) => {
-                            const cellKey = `${rowKey}_${colGroupIndex}_${column}_${columnIndex}`;
-
-                            return (
-                              <td
-                                className={classNames(
-                                  'govuk-table__cell--numeric',
-                                  {
-                                    [styles.borderLeft]:
-                                      columnIndex === 0 && colGroupIndex > 0,
-                                  },
-                                )}
-                                key={cellKey}
-                              >
-                                {column}
-                              </td>
-                            );
-                          }),
-                        )}
-                    </tr>
-                  );
-                })}
-              </tbody>
-            )
-          );
-        })}
-      </table>
-    );
-  },
-);
-
-GroupedDataTable.displayName = 'GroupedDataTable';
 
 const FixedHeaderGroupedDataTable = forwardRef<HTMLElement, Props>(
   (props, ref) => {
@@ -263,7 +106,7 @@ const FixedHeaderGroupedDataTable = forwardRef<HTMLElement, Props>(
             }
           }}
         >
-          <GroupedDataTable
+          <MultiHeaderTable
             {...props}
             className={styles.intersectionTable}
             ref={intersectionTableRef}
@@ -271,14 +114,14 @@ const FixedHeaderGroupedDataTable = forwardRef<HTMLElement, Props>(
             isStickyColumn
             isStickyHeader
           />
-          <GroupedDataTable
+          <MultiHeaderTable
             {...props}
             className={styles.columnTable}
             ref={columnTableRef}
             ariaHidden
             isStickyColumn
           />
-          <GroupedDataTable
+          <MultiHeaderTable
             {...props}
             className={styles.headerTable}
             ref={headerTableRef}
@@ -286,7 +129,7 @@ const FixedHeaderGroupedDataTable = forwardRef<HTMLElement, Props>(
             isStickyHeader
           />
 
-          <GroupedDataTable ref={mainTableRef} {...props} />
+          <MultiHeaderTable ref={mainTableRef} {...props} />
         </div>
       </figure>
     );
