+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/MultiHeaderTable.module.scss
@@ -1,0 +1,23 @@
+@import '~explore-education-statistics-common/src/styles/govuk-base';
+
+$border-width: 4px;
+
+.figure {
+  margin: 0;
+
+  figcaption {
+    padding-bottom: govuk-spacing(4);
+  }
+}
+
+.borderLeft {
+  border-left: $border-width solid govuk-colour('black') !important;
+}
+
+.borderRight {
+  border-right: $border-width solid govuk-colour('black') !important;
+}
+
+.borderBottom {
+  border-bottom: $border-width solid govuk-colour('black') !important;
+}
